+++ lib/sensitive_data_filter/middleware.rb
@@ -1,11 +1,12 @@
 # frozen_string_literal: true
 module SensitiveDataFilter
   module Middleware
+    FILTERABLE = %i(query_params body_params request_params).freeze
   end
 end
 
 require 'sensitive_data_filter/middleware/parameter_parser'
 require 'sensitive_data_filter/middleware/env_parser'
 require 'sensitive_data_filter/middleware/occurrence'
-require 'sensitive_data_filter/middleware/env_filter'
+require 'sensitive_data_filter/middleware/detect'
 require 'sensitive_data_filter/middleware/filter'