+++ lib/sensitive_data_filter/middleware/env_parser.rb
@@ -6,6 +6,7 @@ module SensitiveDataFilter
     class EnvParser
       QUERY_STRING = 'QUERY_STRING'.freeze
       RACK_INPUT   = 'rack.input'.freeze
+      REQUEST_PARAMS = 'action_dispatch.request.request_parameters'.freeze
 
       extend Forwardable
 
@@ -28,6 +29,10 @@ module SensitiveDataFilter
         @parameter_parser.parse(body)
       end
 
+      def request_params
+        @env[REQUEST_PARAMS]
+      end
+
       def query_params=(new_params)
         @env[QUERY_STRING] = Rack::Utils.build_query(new_params)
       end
@@ -36,13 +41,14 @@ module SensitiveDataFilter
         @env[RACK_INPUT] = StringIO.new @parameter_parser.unparse(new_params)
       end
 
-      def copy
-        self.class.new(@env.clone)
+      def request_params=(new_params)
+        @env[REQUEST_PARAMS] = new_params
       end
 
-      def mask!
-        self.query_params = SensitiveDataFilter::Mask.mask(query_params)
-        self.body_params  = SensitiveDataFilter::Mask.mask(body_params)
+      def mutate(mutation)
+        SensitiveDataFilter::Middleware::FILTERABLE.each do |filterable|
+          self.send("#{filterable}=", mutation.send(filterable))
+        end
       end
 
       def_delegators :@request, :ip, :request_method, :url, :content_type, :session