+++ spec/sensitive_data_filter/middleware/filter_spec.rb
@@ -5,35 +5,54 @@ require 'spec_helper'
 require 'sensitive_data_filter/middleware/filter'
 
 describe SensitiveDataFilter::Middleware::Filter do
-  let(:env_filter_class) { double }
-  let(:env_filter)       {
-    double occurrence?: occurrence?, occurrence: occurrence, filtered_env: filtered_env
+  let(:env_parser_class) { double }
+  let(:env_parser)       {
+    double 'EnvParser'
   }
-  let(:occurrence)   { double }
+  let(:occurrence)   { double 'Occurrence' }
+  let(:occurrence_class) { double }
   let(:filtered_env) { double 'filtered_env' }
 
-  let(:app)         { double }
+  let(:app)         { double 'App' }
   let(:middleware)  { SensitiveDataFilter::Middleware::Filter }
   let(:stack)       { middleware.new(app) }
-  let(:env)         { double }
+  let(:env)         { double 'Env' }
+
+  let(:detect) { double }
+  let(:detect_class) { double }
+
+  let(:scan) { double('Scam', matches: []) }
+  let(:changeset) { double('Changeset') }
+
 
   before do
-    stub_const 'SensitiveDataFilter::Middleware::EnvFilter', env_filter_class
-    allow(env_filter_class).to receive(:new).with(env).and_return env_filter
+    stub_const 'SensitiveDataFilter::Middleware::Detect', detect_class
+    stub_const 'SensitiveDataFilter::Middleware::EnvParser', env_parser_class
+
     allow(SensitiveDataFilter).to receive(:handle_occurrence).with occurrence
-    allow(app).to receive(:call).with filtered_env
+    allow(env_parser_class).to receive(:new).and_return(env_parser)
+    allow(detect_class).to receive(:new).with(env_parser).and_return detect
+    allow(detect).to receive(:call).and_return [changeset, scan]
+
+    allow(env_parser).to receive(:mutate).with(changeset)
+
+    stub_const 'SensitiveDataFilter::Middleware::Occurrence', occurrence_class
+    allow(occurrence_class)
+      .to receive(:new).with(env_parser, changeset, scan.matches).and_return occurrence
+
+    allow(app).to receive(:call).with env
     stack.call(env)
   end
 
   context 'when an occurrence is detected' do
-    let(:occurrence?) { true }
+    let(:changeset) { double }
     specify { expect(SensitiveDataFilter).to have_received(:handle_occurrence).with occurrence }
-    specify { expect(app).to have_received(:call).with filtered_env }
+    specify { expect(app).to have_received(:call).with env }
   end
 
   context 'when sensitive data is detected' do
-    let(:occurrence?) { false }
+    let(:changeset) { nil }
     specify { expect(SensitiveDataFilter).not_to have_received(:handle_occurrence) }
-    specify { expect(app).to have_received(:call).with filtered_env }
+    specify { expect(app).to have_received(:call).with env }
   end
 end